feat: a view to show what proposals are assigned to fap reviewers
