+++ apps/e2e/cypress/e2e/FAPs.cy.ts
@@ -1200,6 +1200,76 @@ context('Fap reviews tests', () => {
         '1'
       );
     });
+
+    it('Should be able to see the Reviewers to Assignments view', () => {
+      cy.assignProposalsToFaps({
+        fapInstruments: [
+          { instrumentId: newlyCreatedInstrumentId, fapId: createdFapId },
+        ],
+        proposalPks: [firstCreatedProposalPk],
+      });
+      cy.assignReviewersToFap({
+        fapId: createdFapId,
+        memberIds: [fapMembers.reviewer.id],
+      });
+      cy.assignFapReviewersToProposals({
+        assignments: {
+          memberId: fapMembers.reviewer.id,
+          proposalPk: firstCreatedProposalPk,
+        },
+        fapId: createdFapId,
+      });
+
+      cy.login('officer');
+      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.finishedLoading();
+
+      cy.contains(fapMembers.reviewer.lastName).should('be.visible');
+
+      cy.get('[aria-label="Detail panel visibility toggle"]').first().click();
+      cy.contains(firstCreatedProposalId).should('be.visible');
+    });
+
+    it.only('Should be able to assign proposals to reviewers in the Reviewers to Assignments view', () => {
+      cy.assignProposalsToFaps({
+        fapInstruments: [
+          { instrumentId: newlyCreatedInstrumentId, fapId: createdFapId },
+        ],
+        proposalPks: [firstCreatedProposalPk],
+      });
+      cy.assignReviewersToFap({
+        fapId: createdFapId,
+        memberIds: [fapMembers.reviewer.id],
+      });
+
+      cy.login('officer');
+      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.finishedLoading();
+
+      // cy.get(`[data-cy="assign-reviewer-${fapMembers.reviewer.id}-to-proposal-${firstCreatedProposalPk}"]`).click();
+      cy.contains(fapMembers.reviewer.lastName)
+        .parent()
+        .find('input[type="checkbox"]')
+        .click();
+
+      cy.get('[data-cy="assign-proposals-to-member"]').click();
+
+      cy.get('[role="dialog"]').contains('Proposals to assign to reviewer');
+
+      cy.contains(firstCreatedProposalId)
+        .parent()
+        .find('input[type="checkbox"]')
+        .click();
+
+      cy.get('[data-cy="assign-selected-proposals"]').click();
+
+      clickConfirmOk();
+
+      cy.contains('0 / 1').should('be.visible');
+
+      cy.get('[aria-label="Detail panel visibility toggle"]').first().click();
+      cy.contains(firstCreatedProposalId).should('be.visible');
+    });
   });
 
   describe('Fap Chair role', () => {
@@ -2219,7 +2289,7 @@ context('Fap meeting components tests', () => {
   describe('User Officer role', () => {
     it('Officer should be able to assign proposal to instrument and instrument to call to see it in meeting components', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2248,7 +2318,7 @@ context('Fap meeting components tests', () => {
 
     it('Officer should not be able to submit an instrument if all proposals are not submitted in Fap meetings', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2316,7 +2386,7 @@ context('Fap meeting components tests', () => {
       });
 
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2374,7 +2444,7 @@ context('Fap meeting components tests', () => {
         }
       });
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2529,7 +2599,7 @@ context('Fap meeting components tests', () => {
         }
       });
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2549,7 +2619,7 @@ context('Fap meeting components tests', () => {
 
     it('Officer should be able to see proposals that are marked red if they do not fit in availability time', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2561,7 +2631,7 @@ context('Fap meeting components tests', () => {
 
     it('Officer should be able to update avaliblity time', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2575,7 +2645,7 @@ context('Fap meeting components tests', () => {
 
     it('Officer should be able to edit Fap Meeting form', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2641,7 +2711,7 @@ context('Fap meeting components tests', () => {
       );
 
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2746,7 +2816,7 @@ context('Fap meeting components tests', () => {
       });
 
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2807,7 +2877,7 @@ context('Fap meeting components tests', () => {
       });
 
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2871,7 +2941,7 @@ context('Fap meeting components tests', () => {
       });
 
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2907,7 +2977,7 @@ context('Fap meeting components tests', () => {
 
     it('Officer should be able to set Fap time allocation', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -2959,7 +3029,7 @@ context('Fap meeting components tests', () => {
     it('should use Fap time allocation (if set) when calculating if they fit in available time', () => {
       const newFapTimeAllocation = 15;
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3111,7 +3181,7 @@ context('Fap meeting components tests', () => {
       });
 
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3144,7 +3214,7 @@ context('Fap meeting components tests', () => {
 
     it('Officer should be able to submit an instrument if all proposals Fap meetings are submitted in existing Fap', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3260,7 +3330,7 @@ context('Fap meeting components tests', () => {
         fapId: createdFapId,
       });
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3278,7 +3348,7 @@ context('Fap meeting components tests', () => {
 
     it('Download Fap is working with dialog window showing up', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3515,7 +3585,7 @@ context('Fap meeting components tests', () => {
       cy.changeActiveRole(initialDBData.roles.fapChair);
     });
     it('Fap Chair should be able to edit Fap Meeting form', () => {
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3543,7 +3613,7 @@ context('Fap meeting components tests', () => {
         instrumentId: createdInstrumentId,
         fapId: createdFapId,
       });
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
       cy.get('button[aria-label="Submit instrument"]').should('not.exist');
@@ -3578,7 +3648,7 @@ context('Fap meeting components tests', () => {
 
     it('Fap Chair should be able to update avalibabity time', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3710,7 +3780,7 @@ context('Fap meeting components tests', () => {
 
       cy.login(fapMembers.chair);
       cy.changeActiveRole(initialDBData.roles.fapChair);
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3749,7 +3819,7 @@ context('Fap meeting components tests', () => {
     });
 
     it('Fap Secretary should be able to edit Fap Meeting form', () => {
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3777,7 +3847,7 @@ context('Fap meeting components tests', () => {
         instrumentId: createdInstrumentId,
         fapId: createdFapId,
       });
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
       cy.get('button[aria-label="Submit instrument"]').should('not.exist');
@@ -3812,7 +3882,7 @@ context('Fap meeting components tests', () => {
 
     it('Fap Secretary should be able to update avalibabity time', () => {
       cy.login('officer');
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -3944,7 +4014,7 @@ context('Fap meeting components tests', () => {
 
       cy.login(fapMembers.secretary);
       cy.changeActiveRole(initialDBData.roles.fapSecretary);
-      cy.visit(`/FapPage/${createdFapId}?tab=4`);
+      cy.visit(`/FapPage/${createdFapId}?tab=5`);
 
       cy.finishedLoading();
 
@@ -4690,7 +4760,7 @@ context('Fap meeting exports test', () => {
     });
 
     cy.login('officer');
-    cy.visit('/FapPage/2?tab=4&call=1');
+    cy.visit('/FapPage/2?tab=5&call=1');
 
     cy.get('button[aria-label="Export in Excel"]').click();
     cy.get('[data-cy=preparing-download-dialog').should('not.exist');
@@ -4757,7 +4827,7 @@ context('Fap meeting exports test', () => {
       }
     });
     cy.login('officer');
-    cy.visit('/FapPage/2?tab=4&call=1');
+    cy.visit('/FapPage/2?tab=5&call=1');
 
     cy.get('button[aria-label="Export in Excel"]').click();
     cy.get('[data-cy=preparing-download-dialog').should('not.exist');
@@ -4832,7 +4902,7 @@ context('Fap meeting exports test', () => {
     });
 
     cy.login('officer');
-    cy.visit('/FapPage/2?tab=4&call=1');
+    cy.visit('/FapPage/2?tab=5&call=1');
 
     cy.get('button[aria-label="Export in Excel"]').click();
     cy.get('[data-cy=preparing-download-dialog').should('not.exist');
