+++ apps/e2e/cypress/e2e/generalFaps.cy.ts
@@ -734,7 +734,7 @@ context('General facility access panel tests', () => {
       const fileName1 = 'pdf_5_pages.pdf';
       const fileName2 = 'pdf_3_pages.pdf';
       cy.login('officer');
-      cy.visit(`/FapPage/1?tab=6`);
+      cy.visit(`/FapPage/1?tab=7`);
 
       cy.intercept({
         method: 'POST',
@@ -767,7 +767,7 @@ context('General facility access panel tests', () => {
       cy.contains(fileName2).should('exist');
 
       // Files persist after reload
-      cy.visit(`/FapPage/1?tab=6`);
+      cy.visit(`/FapPage/1?tab=7`);
 
       cy.contains(fileName1).should('exist');
       cy.contains(fileName2).should('exist');
@@ -778,7 +778,7 @@ context('General facility access panel tests', () => {
       cy.contains(fileName2).should('not.exist');
 
       // Files removed after reload
-      cy.visit(`/FapPage/1?tab=6`);
+      cy.visit(`/FapPage/1?tab=7`);
 
       cy.contains(fileName1).should('not.exist');
       cy.contains(fileName2).should('not.exist');
@@ -803,7 +803,7 @@ context('General facility access panel tests', () => {
       const fileName1 = 'pdf_5_pages.pdf';
 
       cy.login('officer');
-      cy.visit(`/FapPage/1?tab=6`);
+      cy.visit(`/FapPage/1?tab=7`);
 
       cy.intercept({
         method: 'POST',
@@ -848,7 +848,7 @@ context('General facility access panel tests', () => {
       const fileName1 = 'pdf_5_pages.pdf';
 
       cy.login('officer');
-      cy.visit(`/FapPage/1?tab=6`);
+      cy.visit(`/FapPage/1?tab=7`);
 
       cy.intercept({
         method: 'POST',
@@ -890,7 +890,7 @@ context('General facility access panel tests', () => {
       const fileName1 = 'pdf_5_pages.pdf';
 
       cy.login('officer');
-      cy.visit(`/FapPage/1?tab=6`);
+      cy.visit(`/FapPage/1?tab=7`);
 
       cy.intercept({
         method: 'POST',
