+++ apps/frontend/src/components/fap/Proposals/MultiRankAssignmentDialog.tsx
@@ -19,7 +19,7 @@ import React, { ChangeEvent } from 'react';
 
 import { BasicUserDetails } from 'generated/sdk';
 
-import { FapAssignedMember } from './AssignFapMemberToProposalModal';
+import { FapAssignedMember } from './ProposalsView/AssignFapMemberToProposalModal';
 
 type MultiRankAssignmentDialogProps = {
   users: BasicUserDetails[];
