+++ apps/frontend/src/components/fap/Proposals/ProposalsView/LegacyFapProposals.tsx
