+++ apps/frontend/src/components/fap/Proposals/ReviewersView/FapReviewersAndAssignmentsView.tsx
@@ -9,24 +9,25 @@ import { useCallsData } from 'hooks/call/useCallsData';
 import { FapProposals } from 'hooks/fap/useFapProposalsData';
 import { useFapInstruments } from 'hooks/instrument/useFapInstruments';
 
-import FapProposalsAndAssignmentsTable from './FapProposalsAndAssignmentsTable';
+import AssignmentsTable from '../AssignmentsTable';
 
-type FapProposalsAndAssignmentsProps = {
+type FapReviewersAndAssignmentsProps = {
   /** Id of the Fap we are assigning members to */
   data: Fap;
   onFapUpdate: (fap: Fap) => void;
   fapProposals: FapProposals;
 };
 
-const FapProposalsAndAssignments = ({
+const FapReviewersAndAssignments = ({
   data: fapData,
   onFapUpdate,
   fapProposals,
-}: FapProposalsAndAssignmentsProps) => {
+}: FapReviewersAndAssignmentsProps) => {
   const { loadingCalls, calls } = useCallsData({
     fapIds: [fapData.id],
     isFapReviewEnded: false,
   });
+
   // NOTE: Default null means load all calls if nothing is selected
   const { loadingInstruments, instruments } = useFapInstruments(
     fapData.id,
@@ -71,15 +72,14 @@ const FapProposalsAndAssignments = ({
           />
         </Grid>
       </Grid>
-      <FapProposalsAndAssignmentsTable
-        data={fapData}
+      <AssignmentsTable
+        fap={fapData}
         onAssignmentsUpdate={onFapUpdate}
-        selectedCallId={call ? +call : null}
-        selectedInstrumentId={instrument ? +instrument : null}
         fapProposals={fapProposals}
+        proposalView={false}
       />
     </>
   );
 };
 
-export default FapProposalsAndAssignments;
+export default FapReviewersAndAssignments;
